refactor: FavoriteVideo 별개의 서비스로 분리
