+++ src/main/java/com/scg/stop/video/service/JobInterviewService.java
@@ -69,26 +69,5 @@ public class JobInterviewService {
         return JobInterviewResponse.from(jobInterview);
     }
 
-    public void createJobInterviewFavorite(Long id, User user) {
-        JobInterview jobInterview = jobInterviewRepository.findById(id).orElseThrow(() ->
-                new BadRequestException(ExceptionCode.ID_NOT_FOUND));
-        FavoriteVideo favoriteVideo = favoriteVideoRepository.findByJobInterviewAndUser(jobInterview, user);
-        if (favoriteVideo != null) {
-            throw new BadRequestException(ExceptionCode.ALREADY_FAVORITE);
-        }
-        FavoriteVideo newFavoriteVideo =  FavoriteVideo.of(jobInterview, user);
-        jobInterview.addFavoriteVideo(newFavoriteVideo);
-        favoriteVideoRepository.save(newFavoriteVideo);
-    }
 
-    public void deleteJobInterviewFavorite(Long id, User user) {
-        JobInterview jobInterview = jobInterviewRepository.findById(id).orElseThrow(() ->
-                new BadRequestException(ExceptionCode.ID_NOT_FOUND));
-        FavoriteVideo favoriteVideo = favoriteVideoRepository.findByJobInterviewAndUser(jobInterview, user);
-        if(favoriteVideo == null) {
-            throw new BadRequestException(ExceptionCode.NOT_FAVORITE);
-        }
-        jobInterview.removeFavoriteVideo(favoriteVideo);
-        favoriteVideoRepository.delete(favoriteVideo);
-    }
 }
