+++ src/main/java/com/scg/stop/video/service/TalkService.java
@@ -86,27 +86,4 @@ public class TalkService {
         talkRepository.delete(talk);
     }
 
-    public void createFavoriteVideo(Long id, User user) {
-        Talk talk = talkRepository.findById(id).orElseThrow(() ->
-                new BadRequestException(ExceptionCode.TALK_ID_NOT_FOUND));
-        FavoriteVideo favoriteVideo = favoriteVideoRepository.findByTalkAndUser(talk, user);
-        if(favoriteVideo != null) {
-            throw new BadRequestException(ExceptionCode.ALREADY_FAVORITE);
-        }
-        FavoriteVideo newFavoriteVideo = FavoriteVideo.of(talk, user);
-        talk.addFavoriteVideo(newFavoriteVideo);
-        favoriteVideoRepository.save(newFavoriteVideo);
-    }
-
-    public void deleteFavoriteVideo(Long id, User user) {
-        Talk talk = talkRepository.findById(id).orElseThrow(() ->
-                new BadRequestException(ExceptionCode.TALK_ID_NOT_FOUND));
-        FavoriteVideo favoriteVideo = favoriteVideoRepository.findByTalkAndUser(talk, user);
-        if(favoriteVideo == null) {
-            throw new BadRequestException(ExceptionCode.NOT_FAVORITE);
-        }
-        talk.removeFavoriteVideo(favoriteVideo);
-        favoriteVideoRepository.delete(favoriteVideo);
-    }
-
 }
